+++ Core/OfficeDevPnP.Core/Pages/ClientSidePageLayoutType.cs
@@ -14,14 +14,16 @@
         /// Home page of modern team sites
         /// </summary>
         Home,
+#if !SP2019
         /// <summary>
         /// Page is an app page, hosting a single SPFX web part full screen
         /// </summary>
         SingleWebPartAppPage,
+#endif
         /// <summary>
         /// Page is a repost / link page
         /// </summary>
         RepostPage
     }
 #endif
-}
+    }